rename clip/polygon to clip/rejoin, to avoid confusion with the forthcoming clipPolygon plugin
