+++ src/clip/rejoin.js
